pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vortex"
version = "0.1.0"
description = "A research-grade optimal control repository for longitudinal vehicle dynamics."
readme = "README.md"
authors = [
  { name = "Research Engineering Team" }
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.23.0",
    "scipy>=1.9.0",
    "casadi>=3.6.0",
    "matplotlib>=3.6.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "flake8"
]

[tool.setuptools.packages.find]
include = ["vortex", "vortex.*"]
